# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Decart AI Camera
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  connected: 'connected',
  closed: 'closed',
} as const;

export interface Room {
  id: string;
  status: RoomStatus;
  createdAt: string;
  mobileConnected: boolean;
  pcConnected: boolean;
}

export interface RoomInput {
  name: string;
}

export type SignalingMessageType = typeof SignalingMessageType[keyof typeof SignalingMessageType];


export const SignalingMessageType = {
  offer: 'offer',
  answer: 'answer',
} as const;

export interface SignalingMessage {
  type: SignalingMessageType;
  sdp: string;
}

export interface SignalingResponse {
  success: boolean;
  message: string;
}

export interface IceCandidate {
  candidate: string;
  sdpMid: string;
  sdpMLineIndex: number;
}

export interface IceCandidatesList {
  candidates: IceCandidate[];
}

export type DecartConfigModel = typeof DecartConfigModel[keyof typeof DecartConfigModel];


export const DecartConfigModel = {
  'lucy-21': 'lucy-2.1',
} as const;

export type DecartConfigMirror = typeof DecartConfigMirror[keyof typeof DecartConfigMirror];


export const DecartConfigMirror = {
  auto: 'auto',
  true: true,
  false: false,
} as const;

export interface DecartConfig {
  apiKey?: string;
  model: DecartConfigModel;
  endpoint: string;
  mirror: DecartConfigMirror;
  enhance: boolean;
  prompt?: string;
}

export type DecartConfigInputModel = typeof DecartConfigInputModel[keyof typeof DecartConfigInputModel];


export const DecartConfigInputModel = {
  'lucy-21': 'lucy-2.1',
} as const;

export type DecartConfigInputMirror = typeof DecartConfigInputMirror[keyof typeof DecartConfigInputMirror];


export const DecartConfigInputMirror = {
  auto: 'auto',
  true: true,
  false: false,
} as const;

export interface DecartConfigInput {
  apiKey: string;
  model?: DecartConfigInputModel;
  mirror?: DecartConfigInputMirror;
  enhance?: boolean;
  prompt: string;
}

export interface Error {
  error: string;
}