# lib/api-zod/src/generated/types/decartConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Decart AI Camera
 * OpenAPI spec version: 0.1.0
 */
import type { DecartConfigMirror } from './decartConfigMirror';
import type { DecartConfigModel } from './decartConfigModel';

export interface DecartConfig {
  apiKey?: string;
  model: DecartConfigModel;
  endpoint: string;
  mirror: DecartConfigMirror;
  enhance: boolean;
  prompt?: string;
}