# lib/api-zod/src/generated/types/decartConfigInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Decart AI Camera
 * OpenAPI spec version: 0.1.0
 */
import type { DecartConfigInputMirror } from './decartConfigInputMirror';
import type { DecartConfigInputModel } from './decartConfigInputModel';

export interface DecartConfigInput {
  apiKey: string;
  model?: DecartConfigInputModel;
  mirror?: DecartConfigInputMirror;
  enhance?: boolean;
  prompt: string;
}