# lib/api-zod/src/generated/types/decartConfigInputModel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Decart AI Camera
 * OpenAPI spec version: 0.1.0
 */

export type DecartConfigInputModel = typeof DecartConfigInputModel[keyof typeof DecartConfigInputModel];


export const DecartConfigInputModel = {
  'lucy-21': 'lucy-2.1',
} as const;