# lib/api-zod/src/generated/types/iceCandidatesList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Decart AI Camera
 * OpenAPI spec version: 0.1.0
 */
import type { IceCandidate } from './iceCandidate';

export interface IceCandidatesList {
  candidates: IceCandidate[];
}