# lib/api-zod/src/generated/types/signalingMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Decart AI Camera
 * OpenAPI spec version: 0.1.0
 */
import type { SignalingMessageType } from './signalingMessageType';

export interface SignalingMessage {
  type: SignalingMessageType;
  sdp: string;
}